+++ tests/oauth2/authorization_code/test_oauth2_authorization_code_async.py
@@ -171,7 +171,7 @@ async def test_oauth2_authorization_code_flow_is_able_to_reuse_client(
         json={
             "access_token": "2YotnFZFEjr1zCsicMWpAA",
             "token_type": "example",
-            "expires_in": 10,
+            "expires_in": 2,
             "example_parameter": "example_value",
         },
         match_content=b"grant_type=authorization_code&redirect_uri=http%3A%2F%2Flocalhost%3A5000%2F&response_type=code&code=SplxlOBeZQQYbYS6WxSbIA",
@@ -190,7 +190,7 @@ async def test_oauth2_authorization_code_flow_is_able_to_reuse_client(
 
     tab.assert_success()
 
-    time.sleep(10)
+    time.sleep(2)
     tab = browser_mock.add_response(
         opened_url="https://provide_code?response_type=code&state=ce9c755b41b5e3c5b64c70598715d5de271023a53f39a67a70215d265d11d2bfb6ef6e9c701701e998e69cbdbf2cee29fd51d2a950aa05f59a20cf4a646099d5&redirect_uri=http%3A%2F%2Flocalhost%3A5000%2F",
         reply_url="http://localhost:5000#code=SplxlOBeZQQYbYS6WxSbIA&state=ce9c755b41b5e3c5b64c70598715d5de271023a53f39a67a70215d265d11d2bfb6ef6e9c701701e998e69cbdbf2cee29fd51d2a950aa05f59a20cf4a646099d5",
@@ -240,7 +240,7 @@ async def test_oauth2_authorization_code_flow_is_able_to_reuse_client_with_token
         json={
             "access_token": "2YotnFZFEjr1zCsicMWpAA",
             "token_type": "example",
-            "expires_in": 10,
+            "expires_in": 2,
             "refresh_token": "tGzv3JOkF0XG5Qx2TlKWIA",
             "example_parameter": "example_value",
         },
@@ -260,7 +260,7 @@ async def test_oauth2_authorization_code_flow_is_able_to_reuse_client_with_token
 
     tab.assert_success()
 
-    time.sleep(10)
+    time.sleep(2)
 
     # response for refresh token grant
     httpx_mock.add_response(